+++ apps/console/src/components/toast/icons/index.ts
@@ -1,0 +1,15 @@
+import AssignmentTurnedIn from './AssignmentTurnedIn'
+import Loading from './Loading'
+import ToastError from './ToastError'
+import ToastInfo from './ToastInfo'
+import ToastSuccess from './ToastSuccess'
+import ToastWarning from './ToastWarning'
+
+export {
+    AssignmentTurnedIn,
+    Loading,
+    ToastError,
+    ToastInfo,
+    ToastSuccess,
+    ToastWarning
+}