+++ packages/@justweb3/ui/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.108 (2025-01-28)
+
+This was a version bump only for @justweb3/ui to align it with other projects, there were no code changes.
+
+## 0.0.107 (2025-01-28)
+
+This was a version bump only for @justweb3/ui to align it with other projects, there were no code changes.
+
+## 0.0.106 (2025-01-28)
+
+This was a version bump only for @justweb3/ui to align it with other projects, there were no code changes.
+
 ## 0.0.105 (2025-01-22)
 
 This was a version bump only for @justweb3/ui to align it with other projects, there were no code changes.